warnings

+++ tests/test_svg_scaling_comprehensive.c
@@ -19,7 +19,7 @@ int main() {
     printf("Scale  | Expected Size | Actual Size | Error%%\n");
     printf("-------|---------------|-------------|--------\n");
     
-    for (i = 0; i < sizeof(test_scales)/sizeof(test_scales[0]); i++) {
+    for (i = 0; i < (int) sizeof(test_scales)/sizeof(test_scales[0]); i++) {
         float scale = test_scales[i];
         V4pPolygonP p = v4p_new(V4P_ABSOLUTE, V4P_WHITE, 0);
         
@@ -68,8 +68,8 @@ int main() {
     
     printf("Scale   | Expected Size | Actual Size | Error%%\n");
     printf("--------|---------------|-------------|--------\n");
-    
-    for (i = 0; i < sizeof(star_scales)/sizeof(star_scales[0]); i++) {
+
+    for (i = 0; i < (int) sizeof(star_scales) / sizeof(star_scales[0]); i++) {
         float scale = star_scales[i];
         V4pPolygonP p = v4p_new(V4P_ABSOLUTE, V4P_WHITE, 0);
         
@@ -103,7 +103,7 @@ int main() {
         
         v4p_destroy(p);
     }
-    
+
     printf("\n");
     
     // Test 3: Fractional scales that were problematic before
@@ -114,8 +114,8 @@ int main() {
     
     printf("Scale  | Expected | Actual | Error%%\n");
     printf("-------|----------|--------|--------\n");
-    
-    for (i = 0; i < sizeof(fractional_scales)/sizeof(fractional_scales[0]); i++) {
+
+    for (i = 0; i < (int) sizeof(fractional_scales) / sizeof(fractional_scales[0]); i++) {
         float scale = fractional_scales[i];
         V4pPolygonP p = v4p_new(V4P_ABSOLUTE, V4P_WHITE, 0);
         
@@ -143,7 +143,7 @@ int main() {
         
         v4p_destroy(p);
     }
-    
+
     printf("\n=== Test Complete ===\n");
     v4p_quit();
     return 0;
